// frontend/utk-workflow/src/utk-ts/src/interfaces.ts
export interface IMasterGrammar {
    variables?: {name: string, value: string}[], 
    components: (IComponent)[],
    knots: IKnot[],
    ex_knots?: IExKnot[], // external precomputed knots
    grid: IGrid,
    grammar: boolean,
    grammar_position?: IComponentPosition
}

export interface IExKnot{
    id: string,
    out_name: string;
    in_name?: string;
    group?: IKnotGroup;
    color_map?: string | IConditionBlock;
    range?: number[];
    domain?: number[];
    scale?: string;
}

/**
 * Interface for map grammar
 */
export interface IMapGrammar {
    variables?: {name: string, value: string}[],
    camera: ICameraData,
    knots: (string | IConditionBlock)[],
    interactions: (InteractionType | IConditionBlock)[],
    plot: {id: string},
    filterKnots?: (number | IConditionBlock)[],
    knotVisibility?: IKnotVisibility[],
    widgets?: IGenericWidget[],
    grammar_type: GrammarType
}

export interface IExternalJoinedJson {
    id: string; // layer id
    incomingId: string[]; // all layers attached to this one
    inValues: number[][][]; // each position store the values for that incoming layer. Inside each position stores a timestep. 
}

/**
 * Interface for plot grammar
 */
export interface IPlotGrammar {
    variables?: {name: string, value: string}[],
    name: string,
    plot: any, // vega-lite spec
    arrangement: string,
    knots: string[],
    interaction: string,
    args?: IPlotArgs,
    interaction_effect?: InteractionEffectType,
    grammar_type: GrammarType
}

export interface IComponent {
    id: string,
    position: IComponentPosition
}

export interface IGrid{
    width: number, // number of horizontal cells
    height: number // number of vertical cells
}

// export interface IToggleKnotsWidget{
//     toggle_knots_widget: {
//         map_id: number,
//         position: IComponentPosition
//     }
// }

export interface IGenericWidget{
    type: WidgetType,
    args?: {categories: ICategory[]}
}

// export interface IGenericWidget{
//     type: WidgetType,
//     map_id?: number, // required to some widgets like toggle knots
//     title?: string,
//     subtitle?: string,
//     categories?: ICategory[], // used with the TOGGLE_KNOT widget
//     position: IComponentPosition
// }

export interface ICategory{
    category_name: string,
    elements: (string | ICategory)[]
}

export interface IComponentPosition{
    width: number[],
    height: number[]
}

export interface IPlotArgs{
    bins?: number | IConditionBlock
}

export interface IKnot {
    id: string,
    group?: IKnotGroup,
    knot_op?: boolean,
    color_map?: string | IConditionBlock,
    integration_scheme: ILinkDescription[],
    range?: number[],
    domain?: number[],
    scale?: string,
}

export interface IKnotGroup{
    group_name: string,
    position: number
}

export interface ILinkDescription {
    spatial_relation?: SpatialRelationType;
    out: {name: string, level: LevelType};
    in?: {name: string, level: LevelType};
    operation: OperationType | IConditionBlock; 
    abstract?: boolean; // if this is a link with an abstract layer
    op?: string; // TODO: the determination if it is a knot operation or not should be done dinamically
    maxDistance?: number; // max distance for the nearest spatial_relation (meters considering that the CRS of the layers is 3395)
    defaultValue?: number; // default value used in Nearest abstract join in case right side is null
}

/**
 * Interface with the camera definition
 */
export interface ICameraData {
    position: number[]; // camera position and look at
    direction: {right: number[], lookAt: number[], up: number[]} // default: right: 0, 0, 1000. lookAt: 0, 0, 0. up: 0, 1, 0
}
// ...
export interface IConditionBlock{
    condition: IConditionElement[],
}

interface IConditionElement{
    test?: string,
    value: any
}

export interface IKnotVisibility{
    knot: string,
    test: string
}